+++ cypress-shared/support/commands.js
@@ -52,11 +52,11 @@ Cypress.Commands.add(
 
       if (selector && contains) {
         cy.scrollTo('top')
-        cy.contains(selector).click()
+        cy.contains(selector).should('be.visible').click()
       } else if (selector) {
-        cy.get(selector).last().click()
+        cy.get(selector).should('be.visible').last().click()
       } else if (contains) {
-        cy.contains(contains).click()
+        cy.contains(contains).should('be.visible').click()
       }
 
       cy.wait(`@${operationName}`, { timeout })
@@ -152,7 +152,7 @@ Cypress.Commands.add('gotoQuickOrder', (b2b = false) => {
   })
 })
 
-Cypress.Commands.add('searchProductinB2B', (product) => {
+Cypress.Commands.add('searchProductinB2B', (product, available = true) => {
   cy.url().then((url) => {
     if (url.includes('checkout')) {
       cy.visit('/')
@@ -167,6 +167,13 @@ Cypress.Commands.add('searchProductinB2B', (product) => {
       .clear()
       .type(product, { force: true })
       .type('{enter}', { force: true })
+    if (available) {
+      cy.get(selectors.searchResult).should('be.visible')
+      cy.get('article div[class*=storefront-permissions-ui]')
+        .should('be.visible')
+        .first()
+        .scrollIntoView()
+    }
   })
 })
 
@@ -205,8 +212,8 @@ Cypress.Commands.add('orderProduct', () => {
       fillContactInfo()
     }
   })
-  cy.get(selectors.PromissoryPayment).click()
-  cy.get(selectors.BuyNowBtn).last().click()
+  cy.get(selectors.PromissoryPayment).should('be.visible').click()
+  cy.get(selectors.BuyNowBtn).last().should('be.visible').click()
   cy.get(selectors.Search, { timeout: 30000 }).should('be.visible')
 })
 
@@ -231,20 +238,21 @@ Cypress.Commands.add('addNewLocation', (country, postalCode, street, city) => {
     .first()
     .clear()
     .should('be.visible')
-    .type(postalCode)
+    .type(postalCode, { delay: 10 })
+  cy.get(selectors.SaveButtonInChangeLocationPopUp).should('be.visible')
   cy.get(selectors.Address)
     .contains('Address Line 1')
     .parent()
     .within(() => {
-      cy.get(selectors.InputText).clear().type(street)
+      cy.get(selectors.InputText).should('be.visible').clear().type(street)
     })
   cy.get(selectors.Address)
     .contains('City')
     .parent()
     .within(() => {
-      cy.get(selectors.InputText).clear().type(city)
+      cy.get(selectors.InputText).should('be.visible').clear().type(city)
     })
-  cy.waitForGraphql('setRegionId', selectors.SaveButton)
+  cy.waitForGraphql('setRegionId', selectors.SaveButtonInChangeLocationPopUp)
   cy.once('uncaught:exception', () => false)
 })