+++ cypress-shared/support/shopper-location/common.js
@@ -3,10 +3,14 @@ import { scroll, updateRetry } from '../common/support'
 import { mockLocation } from './geolocation'
 
 export function verifyShopperLocation() {
+  cy.get(selectors.SaveButtonInChangeLocationPopUp).should('not.exist')
   cy.get(selectors.verifyLocationInHome).should('be.visible')
   // eslint-disable-next-line cypress/no-force
-  cy.get(selectors.AddToCart).contains('Add to cart').click({ force: true })
-  cy.get(selectors.ProceedToCheckOut).click()
+  cy.get(selectors.AddToCart)
+    .contains('Add to cart')
+    .should('be.visible')
+    .click({ force: true })
+  cy.get(selectors.ProceedToCheckOut).should('be.visible').click()
   cy.get(selectors.orderButton).should('be.visible').click()
 }
 
@@ -34,7 +38,9 @@ export function addLocation(data) {
     cy.once('uncaught:exception', () => {
       return false
     })
-    cy.get(selectors.SaveButton).should('be.visible').click()
+    cy.get(selectors.SaveButtonInChangeLocationPopUp)
+      .should('be.visible')
+      .click()
     cy.wait('@setRegionId', { timeout: 10000 })
   })
 }
@@ -78,8 +84,8 @@ export function addAddress(prefix, { address, lat, long }) {
         .clear()
         .type(address.postalCode, { delay: 100 })
       autocomplete(address.city, address.state)
-      cy.get(selectors.saveButton)
-        .find('button')
+      cy.get(selectors.SaveButtonInChangeLocationPopUp)
+        .should('be.visible')
         .click()
         .should(() => {
           expect(localStorage.getItem('orderform')).not.to.be.empty